+++ PowerCMD.sh
@@ -23,7 +23,7 @@ license_owner="neuralpain"
 # link to software website or oss repo
 project_url="https://github.com/neuralpain/qbactivator"
 # basic description of what this script does
-script_description="QuickBooks activation script"
+script_description="QuickBooks POS activation script"
 # change to "true" if your script requires admin
 with_admin=true
 
@@ -45,25 +45,52 @@ pwsh_cache=./cache/pwsh.build.ps1
 # add additional files here
 additional_files=(
   "$bin/ecc/EntClient.dll"
+  "$bin/ecc/EntClientGenuine.dll"
   "$res/doc/instructions.txt"
 )
 # files to exclude in *.min.zip
 exclude_files=(
   "EntClient.dll"
+  "EntClientGenuine.dll"
 )
-# files to exclude in *.min.zip
+# files to cleanup after .zip compression
 remove_files=(
   "instructions.txt"
   "EntClient.dll"
+  "EntClientGenuine.dll"
 )
 # declare a list of your PowerShell functions here
 powershell_functions=(
-  "$functions/Initialize-ObjectsAndVariables.ps1"
-  "$functions/Write-InformationToDisplay.ps1"
-  "$functions/Measure-UserBandwidth.ps1"
-  "$functions/Add-PosLicense.ps1"
-  "$functions/Get-QuickBooksInstaller.ps1"
-  "$functions/Add-ClientPatchModule.ps1"
+  "$functions/ObjectsAndVariables.ps1"
+  # ------------------------------ #
+  # -- add misc scripts -- #
+  "$functions/misc/ClearIntuitData.ps1"
+  "$functions/misc/StopQuickBooksProcesses.ps1"
+  # "$functions/misc/BandwidthTest.ps1"
+  # ------------------------------ #
+  # -- add utility scripts -- #
+  "$functions/utility/Compare-IsValidHash.ps1"
+  "$functions/utility/Format-Text.ps1"
+  "$functions/utility/Invoke-URLInDefaultBrowser.ps1"
+  # "$functions/utility/Measure-UserBandwidth.ps1"
+  "$functions/utility/Show-WebRequestDownloadJobState.ps1"
+  # ------------------------------ #
+  # -- add setup scripts -- #
+  "$functions/setup/PosLicensing.ps1"
+  "$functions/setup/PosActivation.ps1"
+  "$functions/setup/PosInstallation.ps1"
+  "$functions/setup/PosDownload.ps1"
+  "$functions/setup/PosClientModule.ps1"
+  # ------------------------------ #
+  # -- add repair scripts -- #
+  "$functions/repair/ClientModuleRepairs.ps1"
+  # ------------------------------ #
+  # -- add display scripts -- #
+  "$functions/display/DisplayComponents.ps1"
+  "$functions/display/DisplayMenu.ps1"
+  "$functions/display/DisplayLieScreen.ps1"
+  "$functions/display/DisplayError.ps1"
+  # ------------------------------ #
   "$src/Main.ps1"
 )
 
@@ -126,12 +153,16 @@ bundle() {
   echo ":: $script_description" >> $cmd_cache
   echo >> $cmd_cache
   echo "@echo off" >> $cmd_cache
-  echo "@mode 60,18" >> $cmd_cache
+  echo "@mode 60,22" >> $cmd_cache
   echo "@title $script_title v$version" >> $cmd_cache
   add_pwsh
   echo >> $cmd_cache
   # -- add batch code | this is optional -- #
   cat $src/main.cmd >> $cmd_cache
+  echo >> $cmd_cache
+  echo ":qbreadme:" >> $cmd_cache
+  cat $res/doc/instructions.txt >> $cmd_cache
+  echo ":qbreadme:" >> $cmd_cache
   echo >> $cmd_cache
   # -- end batch code -- #
   echo "# ---------- PowerShell Script ---------- #>" >> $cmd_cache